vet-edit-report - move to vet/_lib/components
