+++ src/app/vet/(logged-in)/appointments/submit/[appointmentId]/page.tsx
@@ -4,8 +4,8 @@ import { getAuthenticatedVetActor } from "@/lib/auth";
 import { opFetchAppointment } from "@/lib/bark/operations/op-fetch-appointment";
 import { RoutePath } from "@/lib/route-path";
 import { redirect } from "next/navigation";
-import { SubmitReportForm } from "../../_components/submit-report-form";
 import { APPOINTMENT_STATUS } from "@/lib/bark/models/appointment-status";
+import { SubmitReportForm } from "@/app/vet/_lib/components/submit-report-form";
 
 export default async function Page(props: {
   params: { appointmentId: string };
